+++ src/myPragma.js
@@ -1,18 +1,24 @@
 export default function createElement(tagName, attrs = {}, ...children) {
-    if (tagName === 'null') return children;
+    if (tagName === 'null' || tagName === null) return children;
     if (typeof tagName === 'function') return tagName(attrs, children);
 
     const elem = document.createElement(tagName);
 
     Object.entries(attrs || {}).forEach(([name, value]) => {
-        if (name.startsWith('on') && name.toLowerCase() in window)
-            elem.addEventListener(name.toLowerCase().substr(2), value);
-        else {
-            if (name === 'className')
-                elem.setAttribute('class', value.toString());
-            else if (name === 'htmlFor')
-                elem.setAttribute('for', value.toString());
-            else elem.setAttribute(name, value.toString());
+        if (typeof value !== undefined && value !== null) {
+            if (
+                name.startsWith('on') &&
+                name.toLowerCase() in window &&
+                typeof value === 'function'
+            )
+                elem.addEventListener(name.toLowerCase().substr(2), value);
+            else {
+                if (name === 'className')
+                    elem.setAttribute('class', value.toString());
+                else if (name === 'htmlFor')
+                    elem.setAttribute('for', value.toString());
+                else elem.setAttribute(name, value.toString());
+            }
         }
     });
 
@@ -20,11 +26,8 @@ export default function createElement(tagName, attrs = {}, ...children) {
         if (!child) continue;
         if (Array.isArray(child)) elem.append(...child);
         else {
-            elem.appendChild(
-                child.nodeType === undefined
-                    ? document.createTextNode(child.toString())
-                    : child
-            );
+            if (child.nodeType === undefined) elem.innerHTML += child;
+            else elem.appendChild(child);
         }
     }
     return elem;